+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Firebase v1.0.3 [31-03-2026]
+
+- refactor: update initialize firebase serviceWorker
+- refactor: update firebase-messaging-sw.js
+
 # Firebase v1.0.2 [18-02-2026]
 
 - feat: master updates